% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_check_beast2_packages.R
\name{bbt_check_beast2_packages}
\alias{bbt_check_beast2_packages}
\title{Checks if \code{bbt_run} has the BEAST packages needed to process
its arguments. Will throw if not.}
\usage{
bbt_check_beast2_packages(mcmc, beast2_path = get_default_beast2_bin_path())
}
\arguments{
\item{mcmc}{the MCMC options,
see \link[beautier]{create_mcmc}}

\item{beast2_path}{name of either a BEAST2 binary file
(usually simply \code{beast})
or a BEAST2 jar file
(usually has a \code{.jar} extension).
Use \code{get_default_beast2_bin_path} to get
the default BEAST binary file's path
Use \code{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\description{
Checks if \code{bbt_run} has the BEAST packages needed to process
its arguments. Will throw if not.
}
