% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_create_test_ns_output.R
\name{bbt_create_test_ns_output}
\alias{bbt_create_test_ns_output}
\title{Create testing output similar to when running a BEAST2 run
with nested sampling}
\usage{
bbt_create_test_ns_output()
}
\description{
Create testing output similar to when running a BEAST2 run
with nested sampling
}
\author{
Richèl J.C. Bilderbeek
}
