% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_minimal_run.R
\name{bbt_minimal_run}
\alias{bbt_minimal_run}
\title{Do a minimal run to verify babette works correctly}
\usage{
bbt_minimal_run()
}
\description{
Do a minimal run to verify babette works correctly
}
\author{
Richel J.C. Bilderbeek
}
