% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densitree.R
\name{densitree}
\alias{densitree}
\title{Calls \code{phangorn}'s \link[phangorn]{densiTree} function}
\usage{
densitree(phylos, ...)
}
\description{
Calls \code{phangorn}'s \link[phangorn]{densiTree} function
}
\examples{
  out <- run_beast2(
   get_path("anthus_aco.fas"),
   mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
 )
 densitree(out$anthus_aco_trees)
}
\author{
Richel J.C. Bilderbeek
}
