% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_ids.R
\name{get_alignment_ids}
\alias{get_alignment_ids}
\title{Get the alignment ID from a file with one alignment}
\usage{
get_alignment_ids(xml_filename)
}
\arguments{
\item{xml_filename}{name of a BEAST2 XML input filename}
}
\description{
Get the alignment ID from a file with one alignment
}
\author{
Richel J.C. Bilderbeek
}
