% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_babette_deps.R
\name{install_babette_deps}
\alias{install_babette_deps}
\title{Install the babette dependencies, including BEAST2}
\usage{
install_babette_deps()
}
\description{
Newer code of \code{babette} are not on CRAN yet,
but still on GitHub only. This function installs the GitHub's
master/stable branch.
}
\details{
This function also installs BEAST2 if absent and installs the
\code{NS} BEAST2 package.
}
\author{
Richèl J.C. Bilderbeek
}
