% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_dropbox_path.R
\name{is_dropbox_path}
\alias{is_dropbox_path}
\title{Detect if a path is inside a Dropbox-backed folder}
\usage{
is_dropbox_path(path)
}
\arguments{
\item{path}{A character string representing a file or directory path.}
}
\value{
\code{TRUE} if the path appears to be in a Dropbox folder,
  \code{FALSE} otherwise.
}
\description{
This helper function checks whether a given file or directory path
appears to be located inside a Dropbox-synced folder. It is used to
warn users about potential file-locking issues when running BEAST2
from Dropbox-backed locations.
}
\details{
The detection is based on simple case-insensitive string matching
and is intended as a best-effort heuristic rather than a guarantee.
}
\examples{
is_dropbox_path("~/Dropbox/myproject")
is_dropbox_path("/tmp/myproject")

}
