% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_file_creation.R
\name{prepare_file_creation}
\alias{prepare_file_creation}
\title{Prepare for BEAST2 creating files}
\usage{
prepare_file_creation(inference_model, beast2_options)
}
\description{
The inference model and BEAST2 options contain paths that may point
to sub-sub-sub folders. Create those folders and test
if thes folders can be written to
}
