% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babette.R
\docType{package}
\name{babette}
\alias{babette}
\alias{babette-package}
\title{babette: A package for Bayesian phylogenetics.}
\description{
'babette' provides for an alternative workflow of using
the popular phylogenetics tool BEAST2, including
it peripheral tools. From an alignment and
prior selection, a posterior of jointly estimated
phylogenies and parameter estimates is generated.
}
\seealso{
These are packages associated with 'beautier':
\itemize{
  \item{
    The package 'beautier' can create
    BEAST2 input files from R
  }
  \item{
    The package 'beastier' can run
    BEAST2 from R
  }
  \item{
    The package 'tracerer' can parse
    BEAST2 output files from R
  }
}
}
