% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babette.R
\docType{package}
\name{babette}
\alias{babette}
\alias{babette-package}
\title{babette: A package for Bayesian phylogenetics.}
\description{
'babette' provides for an alternative workflow of using
the popular phylogenetics tool BEAST2, including
it peripheral tools. From an alignment and
prior selection, a posterior of jointly estimated
phylogenies and parameter estimates is generated.
}
\section{babette subpackages}{


'babette' is mostly a glue between three subpackages:
  * \link[beautier](beautier)
  * beastier
  * tracerer
}

