% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babette.R
\docType{package}
\name{babette}
\alias{babette}
\title{babette: A package for Bayesian phylogenetics.}
\description{
'babette' provides for an alternative workflow of using
the popular phylogenetics tool 'BEAST2', including
it peripheral tools. From an alignment and
prior selection, a posterior of jointly estimated
phylogenies and parameter estimates is generated.
}
\seealso{
These are packages associated with 'beautier':
\itemize{
  \item{
    The package '\link[beautier]{beautier}' can create
    BEAST2 input files from R
  }
  \item{
    The package '\link[beastier]{beastier}' can run
    BEAST2 from R
  }
  \item{
    The package '\link[mauricer]{mauricer}' can install
    BEAST2 packages from R
  }
  \item{
    The package '\link[tracerer]{tracerer}' can parse
    BEAST2 output files from R
  }
}
}
