% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbt_self_test.R
\name{bbt_self_test}
\alias{bbt_self_test}
\title{Do a self test to verify babette works correctly}
\usage{
bbt_self_test()
}
\description{
Do a self test to verify babette works correctly
}
\author{
Richel J.C. Bilderbeek
}
