% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_bbt_run_output.R
\name{create_test_bbt_run_output}
\alias{create_test_bbt_run_output}
\title{Get an example output of \code{\link{bbt_run}}
or \code{\link{bbt_run_from_model}}.}
\usage{
create_test_bbt_run_output()
}
\value{
the same results as \code{\link{bbt_run}}
  or \code{\link{bbt_run_from_model}}
}
\description{
This output is used in testing.
}
\examples{
beastier::remove_beaustier_folder()
beastier::check_empty_beaustier_folders()

create_test_bbt_run_output()

beastier::remove_beaustier_folder()
beastier::check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
