% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_babette_paths.R
\name{get_babette_paths}
\alias{get_babette_paths}
\title{Get the full paths of files in the \code{inst/extdata} folder}
\usage{
get_babette_paths(filenames)
}
\arguments{
\item{filenames}{the files' names, without the path}
}
\value{
the filenames' full paths, if and only if
  all files are present. Will stop otherwise.
}
\description{
Get the full paths of files in the \code{inst/extdata} folder
}
\examples{
beastier::remove_beaustier_folder()
beastier::check_empty_beaustier_folders()

get_babette_paths(c("anthus_aco.fas", "anthus_nd2.fas"))

beastier::remove_beaustier_folder()
beastier::check_empty_beaustier_folders()
}
\seealso{
for one file, use \code{\link{get_babette_path}}
}
\author{
Richèl J.C. Bilderbeek
}
