% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_beast2_output_to_ns.R
\name{parse_beast2_output_to_ns}
\alias{parse_beast2_output_to_ns}
\title{Parse BEAST2 NS output}
\usage{
parse_beast2_output_to_ns(output)
}
\arguments{
\item{output}{screen output}
}
\value{
a list with the following elements:
  \itemize{
    \item \code{marg_log_lik} the marginal log likelihood estimate
    \item \code{marg_log_lik_sd} the standard deviation around the estimate
  }
}
\description{
Parse the BEAST2 output when run with the BEAST2 NS ('Nested Sampling')
package.
}
\examples{
beastier::remove_beaustier_folder()
beastier::check_empty_beaustier_folders()

parse_beast2_output_to_ns(
  output = create_test_ns_output()
)

beastier::remove_beaustier_folder()
beastier::check_empty_beaustier_folders()
}
\seealso{
use \code{\link{create_test_ns_output}} to obtain
a test screen output.
}
\author{
Richèl J.C. Bilderbeek
}
