% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densitree.R
\name{plot_densitree}
\alias{plot_densitree}
\title{Draw multiple trees on top of one another.}
\usage{
plot_densitree(phylos, library = "ggtree", ...)
}
\arguments{
\item{phylos}{one or more phylogenies, must be of class \code{multiPhylo}}

\item{library}{the library to use:
\itemize{
  \item \code{ggtree}: use \code{ggtree}'s
    \link[ggtree]{ggtree} function
  \item \code{phangorn}: use \code{phangorn}'s
    \link[phangorn]{densiTree} function
}}

\item{...}{options to be passed to \code{phangorn}'s
\link[phangorn]{densiTree} function}
}
\value{
nothing. Will produce a plot.
}
\description{
Draw multiple trees on top of one another.
}
\examples{
  out <- bbt_run(
   get_babette_path("anthus_aco.fas"),
   mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
 )
 plot_densitree(out$anthus_aco_trees)
}
\author{
Richel J.C. Bilderbeek
}
