% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densitree.R
\name{plot_densitree}
\alias{plot_densitree}
\title{Calls \code{phangorn}'s \link[phangorn]{densiTree} function}
\usage{
plot_densitree(phylos, ...)
}
\arguments{
\item{phylos}{one or more phylogenies, must be of class \code{multiPhylo}}

\item{...}{options to be passed to \code{phangorn}'s
\link[phangorn]{densiTree} function}
}
\description{
Calls \code{phangorn}'s \link[phangorn]{densiTree} function
}
\examples{
  out <- bbt_run(
   get_babette_path("anthus_aco.fas"),
   mcmc = create_mcmc(chain_length = 10000, store_every = 1000)
 )
 plot_densitree(out$anthus_aco_trees)
}
\author{
Richel J.C. Bilderbeek
}
