% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_babette.R
\name{update_babette}
\alias{update_babette}
\title{Deprecated function.}
\usage{
update_babette(upgrade = "default")
}
\arguments{
\item{upgrade}{Deprecated.}
}
\value{
Nothing.
}
\description{
Update all babette dependencies, by installing their
latest versions.
}
\details{
See \url{https://github.com/richelbilderbeek/babetteinstall}
how to do this.
}
\author{
Giovanni Laudanno, Richèl J.C. Bilderbeek
}
